Clientside permission-check for delete Project

+++ source/js/app/controller/projects.js
@@ -10,17 +10,25 @@ App.ProjectOverviewController = Ember.ObjectController.extend({
 	projectDeleteCommitted : false,
 
 	deleteProject: function(projectToDelete) {
+		console.log(App.get("session.sessionUserId"));
+		console.log(projectToDelete.get('owner.id'));
+
+		if (App.get("session.sessionUserId")==projectToDelete.get('owner.id')){
 			this.projectDeleteCommitted = false;
 
 			var confirmResult = confirm("Delete project "+projectToDelete.get('title')+" ?");
 
-			if (confirmResult){
+			if(confirmResult){
 				this.projectDeleteCommitted = true;
 				var project = projectToDelete;
 				showFlashMessageFor(project);
 				project.deleteRecord();
 				App.store.commit();
 			}
+		}
+		else{
+			alert("You have no permission to delete this project");
+		}
 	}
 });
